# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeResumeRequest {
  /** Extracted text from the resume PDF */
  resumeText: string;
  /** The job role id to match against (use "__custom__" for custom roles) */
  jobRole: string;
  /** Required skills for a custom job role (only used when jobRole is "__custom__") */
  customSkills?: string[];
  /** Title for the custom job role */
  customRoleTitle?: string;
}

export interface AnalyzeResumeResponse {
  candidateName?: string;
  email?: string;
  phone?: string;
  /** Overall match score 0-100 */
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  projectScore: number;
  detectedSkills: string[];
  missingSkills: string[];
  matchedSkills: string[];
  requiredSkills: string[];
  education: string[];
  experience: string;
  projects: string[];
  certifications: string[];
  resumeSummary: string;
  strengthAreas: string[];
  improvementTips: string[];
  skillRecommendations: string[];
  explainability: string;
  skillCoveragePercent: number;
}

export interface JobRole {
  id: string;
  title: string;
  category: string;
  requiredSkills: string[];
  description: string;
}

export interface JobRolesResponse {
  roles: JobRole[];
  categories: string[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}