# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Accepts PDF resume text and job role, returns detailed analysis
 * @summary Analyze a resume against a job role
 */
export const AnalyzeResumeBody = zod.object({
  resumeText: zod.string().describe("Extracted text from the resume PDF"),
  jobRole: zod
    .string()
    .describe(
      'The job role id to match against (use \"__custom__\" for custom roles)',
    ),
  customSkills: zod
    .array(zod.string())
    .optional()
    .describe(
      'Required skills for a custom job role (only used when jobRole is \"__custom__\")',
    ),
  customRoleTitle: zod
    .string()
    .optional()
    .describe("Title for the custom job role"),
});

export const AnalyzeResumeResponse = zod.object({
  candidateName: zod.string().optional(),
  email: zod.string().optional(),
  phone: zod.string().optional(),
  overallScore: zod.number().describe("Overall match score 0-100"),
  skillScore: zod.number(),
  experienceScore: zod.number(),
  educationScore: zod.number(),
  projectScore: zod.number(),
  detectedSkills: zod.array(zod.string()),
  missingSkills: zod.array(zod.string()),
  matchedSkills: zod.array(zod.string()),
  requiredSkills: zod.array(zod.string()),
  education: zod.array(zod.string()),
  experience: zod.string(),
  projects: zod.array(zod.string()),
  certifications: zod.array(zod.string()),
  resumeSummary: zod.string(),
  strengthAreas: zod.array(zod.string()),
  improvementTips: zod.array(zod.string()),
  skillRecommendations: zod.array(zod.string()),
  explainability: zod.string(),
  skillCoveragePercent: zod.number(),
});

/**
 * Returns list of available job roles with their required skills and categories
 * @summary Get all predefined job roles grouped by category
 */
export const GetJobRolesResponse = zod.object({
  roles: zod.array(
    zod.object({
      id: zod.string(),
      title: zod.string(),
      category: zod.string(),
      requiredSkills: zod.array(zod.string()),
      description: zod.string(),
    }),
  ),
  categories: zod.array(zod.string()),
});