# lib/api-zod/src/generated/types/analyzeResumeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeResumeRequest {
  /** Extracted text from the resume PDF */
  resumeText: string;
  /** The job role id to match against (use "__custom__" for custom roles) */
  jobRole: string;
  /** Required skills for a custom job role (only used when jobRole is "__custom__") */
  customSkills?: string[];
  /** Title for the custom job role */
  customRoleTitle?: string;
}