# lib/api-zod/src/generated/types/analyzeResumeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeResumeResponse {
  candidateName?: string;
  email?: string;
  phone?: string;
  /** Overall match score 0-100 */
  overallScore: number;
  skillScore: number;
  experienceScore: number;
  educationScore: number;
  projectScore: number;
  detectedSkills: string[];
  missingSkills: string[];
  matchedSkills: string[];
  requiredSkills: string[];
  education: string[];
  experience: string;
  projects: string[];
  certifications: string[];
  resumeSummary: string;
  strengthAreas: string[];
  improvementTips: string[];
  skillRecommendations: string[];
  explainability: string;
  skillCoveragePercent: number;
}