# lib/api-zod/src/generated/types/jobRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface JobRole {
  id: string;
  title: string;
  category: string;
  requiredSkills: string[];
  description: string;
}