# lib/api-zod/src/generated/types/jobRolesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Powered Resume Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { JobRole } from "./jobRole";

export interface JobRolesResponse {
  roles: JobRole[];
  categories: string[];
}